deployment/requirements.txt
fastapi==0.125.0
uvicorn==0.40.0
pydantic==2.12.5
pandas==2.2.3
joblib==1.5.3
scikit-learn==1.3.0
scipy==1.15.1
numpy==1.26.4
evidently==0.4.15
python-multipart==0.0.9
pymongo==4.6.1
motor==3.3.2
python-dotenv==1.0.1
